requirements.txt
matplotlib==3.5.2
numpy == 1.23.1
pytorch == 1.12.1(cuda 11.6)
python == 3.8.0
scipy  == 1.8.1
torchaudio == 0.12.1       
torchvision == 0.13.1